refactor(dotenv): move `parse()` to own file
