+++ dotenv/mod.ts
@@ -174,6 +174,10 @@
  * @module
  */
 
+import { parse } from "./parse.ts";
+
+export * from "./parse.ts";
+
 export interface LoadOptions {
   /**
    * Optional path to `.env` file. To prevent the default value from being
@@ -221,51 +225,6 @@ export interface LoadOptions {
    * @default {"./.env.defaults"}
    */
   defaultsPath?: string | null;
-}
-
-type LineParseResult = {
-  key: string;
-  unquoted: string;
-  interpolated: string;
-  notInterpolated: string;
-};
-
-type CharactersMap = { [key: string]: string };
-
-const RE_KeyValue =
-  /^\s*(?:export\s+)?(?<key>[a-zA-Z_]+[a-zA-Z0-9_]*?)\s*=[\ \t]*('\n?(?<notInterpolated>(.|\n)*?)\n?'|"\n?(?<interpolated>(.|\n)*?)\n?"|(?<unquoted>[^\n#]*)) *#*.*$/gm;
-
-const RE_ExpandValue =
-  /(\${(?<inBrackets>.+?)(\:-(?<inBracketsDefault>.+))?}|(?<!\\)\$(?<notInBrackets>\w+)(\:-(?<notInBracketsDefault>.+))?)/g;
-
-export function parse(rawDotenv: string): Record<string, string> {
-  const env: Record<string, string> = {};
-
-  let match;
-  const keysForExpandCheck = [];
-
-  while ((match = RE_KeyValue.exec(rawDotenv)) !== null) {
-    const { key, interpolated, notInterpolated, unquoted } = match
-      ?.groups as LineParseResult;
-
-    if (unquoted) {
-      keysForExpandCheck.push(key);
-    }
-
-    env[key] = typeof notInterpolated === "string"
-      ? notInterpolated
-      : typeof interpolated === "string"
-      ? expandCharacters(interpolated)
-      : unquoted.trim();
-  }
-
-  //https://github.com/motdotla/dotenv-expand/blob/ed5fea5bf517a09fd743ce2c63150e88c8a5f6d1/lib/main.js#L23
-  const variablesMap = { ...env };
-  keysForExpandCheck.forEach((key) => {
-    env[key] = expand(env[key], variablesMap);
-  });
-
-  return env;
 }
 
 export function loadSync(
@@ -360,19 +319,6 @@ async function parseFile(
   }
 }
 
-function expandCharacters(str: string): string {
-  const charactersMap: CharactersMap = {
-    "\\n": "\n",
-    "\\r": "\r",
-    "\\t": "\t",
-  };
-
-  return str.replace(
-    /\\([nrt])/g,
-    ($1: keyof CharactersMap): string => charactersMap[$1],
-  );
-}
-
 function assertSafe(
   conf: Record<string, string>,
   confExample: Record<string, string>,
@@ -420,32 +366,6 @@ export class MissingEnvVarsError extends Error {
     this.name = "MissingEnvVarsError";
     this.missing = missing;
     Object.setPrototypeOf(this, new.target.prototype);
-  }
-}
-
-function expand(str: string, variablesMap: { [key: string]: string }): string {
-  if (RE_ExpandValue.test(str)) {
-    return expand(
-      str.replace(RE_ExpandValue, function (...params) {
-        const {
-          inBrackets,
-          inBracketsDefault,
-          notInBrackets,
-          notInBracketsDefault,
-        } = params[params.length - 1];
-        const expandValue = inBrackets || notInBrackets;
-        const defaultValue = inBracketsDefault || notInBracketsDefault;
-
-        let value: string | undefined = variablesMap[expandValue];
-        if (value === undefined) {
-          value = Deno.env.get(expandValue);
-        }
-        return value === undefined ? expand(defaultValue, variablesMap) : value;
-      }),
-      variablesMap,
-    );
-  } else {
-    return str;
   }
 }
